rename LTX2 generation backend.
