+++ apps/dreamverse/arch.md
@@ -145,7 +145,7 @@ Model generation has a separate ownership boundary inside each GPU process:
 
 - `apps/dreamverse/dreamverse/generation_worker.py` selects the backend that the active model profile declares and owns
   the backend lifecycle.
-- `apps/dreamverse/dreamverse/video_generation.py` owns LTX-2 generator configuration, video and audio continuation, and
+- `apps/dreamverse/dreamverse/ltx2_generation.py` owns LTX-2 generator configuration, video and audio continuation, and
   runtime LoRA application.
 - `apps/dreamverse/dreamverse/minimax_h3_generation.py` owns the VSA data-free FastH3 adapter, FastH3 generator and
   request configuration, and last-frame continuation through MiniMax H3 first-frame conditioning.
